Correction de code caca && déplacement des fonctions dans les bons fichiers

+++ libft/strnstr.c
@@ -6,7 +6,7 @@
 /*   By: edescoin <edescoin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/07 13:19:48 by edescoin          #+#    #+#             */
-/*   Updated: 2016/11/08 16:35:19 by edescoin         ###   ########.fr       */
+/*   Updated: 2016/11/08 19:15:21 by edescoin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
 	long	i;
-	long	j;
+	size_t	j;
 	char	*tmp;
 
 	if (little[0] == '\0')
